requirements.txt
# Core dependencies from pyproject.toml
bleach[css]~=6.2
celery[redis]==5.5.3
crispy-bootstrap5==2025.6
django==5.2.7
django-activity-stream==2.0.0
django-axes[ipware]==8.0.0
django-bootstrap-breadcrumbs2==1.0.0
django-compressor==4.5.1
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-email-verification==0.3.3
django-environ==0.12.0
django-filter==25.1
django-formtools~=2.5
django-prometheus==2.4.1
django-recaptcha==4.1.0
django-redis==6.0.0
django-simple-history==3.10.1
django-sortedm2m==4.0.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework-simplejwt[crypto]==5.5.1
drf-spectacular[sidecar]==0.28.0
easy-thumbnails==2.10.1
flower==2.0.1
fontawesomefree~=6.6.0
icalendar==6.3.1
invoke==2.2.0
lingua-language-detector==2.1.1
openfoodfacts==3.1.0
packaging==25.0
pillow==11.3.0
psycopg==3.2.10
reportlab==4.4.4
requests==2.32.5
tqdm==4.67.1
tzdata==2025.2

# Railway-specific dependencies
gunicorn==23.0.0
whitenoise==6.8.2
django-redis==6.0.0
